+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * reorganized station functions and names to match grid functionality
 * made url and query generation more consistent
-* added possibility to download grids
+* added possibility to download grids (daily and monthly)
 
 
 # rzamg 0.1.0